Moved Dialogs to UI Components
